feat1207: 新增商品增删及购买的接口

+++ src/main/java/com/example/walletpayment/mybatis/entity/BankAccount.java
@@ -13,7 +13,7 @@ public class BankAccount {
 
     private String bankName;
 
-    private Integer balance;
+    private Double balance;
 
     private String accountType;
